rename denryoku to keisoku
