+++ Cargo.toml
@@ -18,7 +18,7 @@ members = [
     "crates/cli-storage",
     "crates/cli-tools",
     "crates/test-runner",
-    "crates/denryoku",
+    "crates/keisoku",
     "crates/kanka",
 ]
 resolver = "3"
@@ -59,7 +59,7 @@ nagare = { path = "crates/nagare", default-features = false }
 download-manager = { path = "crates/download-manager", default-features = false }
 mock-registry = { path = "crates/mock-registry", default-features = false }
 benchmarks = { path = "crates/benchmarks", default-features = false }
-denryoku = { path = "crates/denryoku", default-features = false }
+keisoku = { path = "crates/keisoku", default-features = false }
 kanka = { path = "crates/kanka", default-features = false }
 uzu = { path = "crates/uzu", default-features = false }
 xgrammar = { package = "xgrammar-rs", git = "https://github.com/trymirai/xgrammar-rs.git", tag = "0.2.2", features = ["tokenizers"] }
@@ -75,6 +75,7 @@ tokenizers = { version = "0.22", features = [
 objc2 = "0.6.4"
 objc2-foundation = "0.3.2"
 objc2-core-foundation = "0.3.2"
+objc2-io-kit = { version = "0.3.2", default-features = false, features = ["libc"] }
 objc2-security = "0.3.2"
 obfstr = "0.4.4"
 metal = { package = "mtl-rs", version = "0.1.12" }
@@ -214,7 +215,6 @@ rodio = { version = "0.22.2", default-features = false, features = [
     "playback",
     "wav",
 ] }
-macmon = "0.7"
 
 [workspace.lints.clippy]
 # TODO: add large_stack_frames, large_stack_arrays, trivially_copy_pass_by_ref, needless_pass_by_value, map_err_ignore, use_self, cloned_instead_of_copied, manual_let_else, semicolon_if_nothing_returned, uninlined_format_args, lossy_float_literal
