+++ crates/keisoku/src/component.rs
@@ -1,6 +1,8 @@
 use core::fmt;
 
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, Serialize, Deserialize)]
 pub enum Component {
     Cpu,
     Gpu,
